lab04/requirements.txt
altair==4.1.0; python_version >= "3.6"
appnope==0.1.2; platform_system == "Darwin" and python_version >= "3.7" and sys_platform == "darwin"
argon2-cffi==20.1.0; python_version >= "3.6"
astor==0.8.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
async-generator==1.10; python_full_version >= "3.6.1" and python_version >= "3.6"
attrs==20.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
backcall==0.2.0; python_version >= "3.7"
base58==2.1.0; python_version >= "3.6"
bleach==3.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
blinker==1.4; python_version >= "3.6"
cachetools==4.2.1; python_version >= "3.6" and python_version < "4.0"
certifi==2020.12.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
cffi==1.14.5; implementation_name == "pypy" and python_version >= "3.6"
chardet==4.0.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
click==7.1.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0"
decorator==4.4.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.2.0" and python_version >= "3.7"
defusedxml==0.7.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
entrypoints==0.3; python_version >= "3.6"
gitdb==4.0.6; python_version >= "3.6"
gitpython==3.1.14; python_version >= "3.6"
idna==2.10; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
importlib-metadata==3.7.3; python_version < "3.8" and python_version >= "3.6"
ipykernel==5.5.0; python_version >= "3.6"
ipython-genutils==0.2.0; python_version >= "3.7"
ipython==7.21.0; python_version >= "3.7"
ipywidgets==7.6.3; python_version >= "3.6"
jedi==0.18.0; python_version >= "3.7"
jinja2==2.11.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
jsonschema==3.2.0; python_version >= "3.6"
jupyter-client==6.1.12; python_full_version >= "3.6.1" and python_version >= "3.6"
jupyter-core==4.7.1; python_version >= "3.6"
jupyterlab-pygments==0.1.2; python_version >= "3.6"
jupyterlab-widgets==1.0.0; python_version >= "3.6"
markupsafe==1.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
mistune==0.8.4; python_version >= "3.6"
nbclient==0.5.3; python_full_version >= "3.6.1" and python_version >= "3.6"
nbconvert==6.0.7; python_version >= "3.6"
nbformat==5.1.2; python_full_version >= "3.6.1" and python_version >= "3.6"
nest-asyncio==1.5.1; python_full_version >= "3.6.1" and python_version >= "3.6"
notebook==6.3.0; python_version >= "3.6"
numpy==1.20.1; python_full_version >= "3.6.1" and python_version >= "3.7" and python_version < "3.9"
packaging==20.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pandas==1.1.5; python_full_version >= "3.6.1" and python_version >= "3.6"
pandocfilters==1.4.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
parso==0.8.1; python_version >= "3.7"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.7"
pickleshare==0.7.5; python_version >= "3.7"
pillow==8.1.2; python_version >= "3.6"
prometheus-client==0.9.0; python_version >= "3.6"
prompt-toolkit==3.0.18; python_full_version >= "3.6.1" and python_version >= "3.7"
protobuf==3.15.6; python_version >= "3.6"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.7" and os_name != "nt"
py==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" and implementation_name == "pypy" or implementation_name == "pypy" and python_version >= "3.6" and python_full_version >= "3.4.0"
pyarrow==3.0.0; python_version < "3.9" and python_version >= "3.6"
pycparser==2.20; python_version >= "3.6" and python_full_version < "3.0.0" and implementation_name == "pypy" or implementation_name == "pypy" and python_version >= "3.6" and python_full_version >= "3.4.0"
pydeck==0.6.1; python_version >= "3.6"
pygments==2.8.1; python_version >= "3.7"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyrsistent==0.17.3; python_version >= "3.6"
python-dateutil==2.8.1; python_full_version >= "3.6.1" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
pytz==2021.1; python_full_version >= "3.6.1" and python_version >= "3.6"
pywin32==300; sys_platform == "win32" and python_version >= "3.6"
pywinpty==0.5.7; os_name == "nt" and python_version >= "3.6"
pyzmq==22.0.3; python_version >= "3.6"
requests==2.25.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
send2trash==1.5.0; python_version >= "3.6"
six==1.15.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
smmap==3.0.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
streamlit==0.79.0; python_version >= "3.6"
terminado==0.9.3; python_version >= "3.6"
testpath==0.4.4; python_version >= "3.6"
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
toolz==0.11.1; python_version >= "3.6"
tornado==6.1; python_version >= "3.6"
traitlets==5.0.5; python_full_version >= "3.6.1" and python_version >= "3.7"
typing-extensions==3.7.4.3; python_version < "3.8" and python_version >= "3.6"
tzlocal==2.1; python_version >= "3.6"
urllib3==1.26.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
validators==0.18.2; python_version >= "3.6"
watchdog==2.0.2; platform_system != "Darwin" and python_version >= "3.6"
wcwidth==0.2.5; python_full_version >= "3.6.1" and python_version >= "3.7"
webencodings==0.5.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
widgetsnbextension==3.5.1; python_version >= "3.6"
zipp==3.4.1; python_version < "3.8" and python_version >= "3.6"
